% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nastawZegar.R
\name{nastawZegar}
\alias{nastawZegar}
\title{Nastawienie zegara, który uruchamia maszynę uśmiercającą lub nieuśmiercającą baranka.}
\usage{
nastawZegar(pudelko)
}
\arguments{
\item{pudelko}{pudelko. Pudelko z barankiem.}
}
\description{
Nastawienie zegara, który uruchamia maszynę uśmiercającą lub nieuśmiercającą baranka.
}
