% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otworz.R
\name{otworz}
\alias{otworz}
\title{Funkcja otwierająca pudełko. Pokazuje co jest wewnątrz.}
\usage{
otworz(pudelko)
}
\arguments{
\item{pudelko}{pudelko. Pudelko z barankiem.}
}
\value{
NULL

otworz(pudelko)
}
\description{
Funkcja otwierająca pudełko. Pokazuje co jest wewnątrz.
}
