% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/potrzasnij.R
\name{potrzasnij}
\alias{potrzasnij}
\title{Funkcja trzęsąca pudełkiem.}
\usage{
potrzasnij(pudelko)
}
\arguments{
\item{pudelko}{pudelko. Pudelko z barankiem.}
}
\value{
NULL

potrzasnij(pudelko)
}
\description{
Funkcja trzęsąca pudełkiem.
}
