% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.baranek.R
\name{print.baranek}
\alias{print.baranek}
\title{Funkcja wyświetlająca baranka. Lepiej jednak go nie wyświetlać...}
\usage{
\method{print}{baranek}(baranek)
}
\arguments{
\item{baranek}{baranek}
}
\value{
NULL

print(baranek)
}
\description{
Funkcja wyświetlająca baranka. Lepiej jednak go nie wyświetlać...
}
