% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wsadz.R
\name{wsadz}
\alias{wsadz}
\title{Funkcja, która wsadza rzeczy (np. baranki) do pudełka.}
\usage{
wsadz(pudelko, baranek)
}
\arguments{
\item{pudelko}{pudelko. Pudelko z barankiem.}

\item{baranek}{baranek (albo coś innego).}
}
\value{
NULL

wsadz(pudelko, baranek)
}
\description{
Funkcja, która wsadza rzeczy (np. baranki) do pudełka.
}
